refactor: simplify multiplex hash logging cast

+++ src/projects/providers/multiplex/multiplex_application.cpp
@@ -198,7 +198,7 @@ namespace pvd
                 continue;
             }
 
-            logtt("Found multiplex file : %s, mtime : %ld, hash : %zu", multiplex_file_info._file_path.CStr(), static_cast<long>(multiplex_file_info._file_stat.st_mtime), static_cast<size_t>(multiplex_file_info._file_path.Hash()));
+            logtt("Found multiplex file : %s, mtime : %" PRIdMAX ", hash : %zu", multiplex_file_info._file_path.CStr(), static_cast<intmax_t>(multiplex_file_info._file_stat.st_mtime), multiplex_file_info._file_path.Hash());
 
             multiplex_files.push_back(multiplex_file_info);
         }
